+++ packages/collections/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/language-collections
 
+## 0.9.4 - 03 September 2026
+
+### Patch Changes
+
+- fd1b2be: Security update
+- Updated dependencies \[654026d]
+- Updated dependencies \[fd1b2be]
+  - @openfn/language-common@3.3.5
+
 ## 0.9.3 - 30 June 2026
 
 ### Patch Changes