+++ packages/fhir-opensrp/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @openfn/language-fhir-opensrp
 
+## 2.0.1 - 03 September 2026
+
+### Patch Changes
+
+- fd1b2be: Security update
+- Updated dependencies \[654026d]
+- Updated dependencies \[96c6822]
+- Updated dependencies \[fd1b2be]
+  - @openfn/language-common@3.3.5
+  - @openfn/language-fhir-4@0.6.0
+
 ## 2.0.0 - 18 August 2026
 
 ### Major Changes