+++ packages/godata/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/language-godata
 
+## 3.6.4 - 03 September 2026
+
+### Patch Changes
+
+- 19f1642: Security update
+- Updated dependencies \[654026d]
+- Updated dependencies \[fd1b2be]
+  - @openfn/language-common@3.3.5
+
 ## 3.6.3 - 28 July 2026
 
 ### Patch Changes