+++ packages/image-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/language-image-fn
 
+## 1.0.2 - 03 September 2026
+
+### Patch Changes
+
+- Updated dependencies \[654026d]
+- Updated dependencies \[fd1b2be]
+  - @openfn/language-common@3.3.5
+
 ## 1.0.1 - 13 July 2026
 
 ### Patch Changes