+++ packages/langfuse/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/language-langfuse
 
+## 1.0.1 - 03 September 2026
+
+### Patch Changes
+
+- fd1b2be: Security update
+- Updated dependencies \[654026d]
+- Updated dependencies \[fd1b2be]
+  - @openfn/language-common@3.3.5
+
 ## 1.0.0 - 28 July 2026
 
 Initial release with langfuse operation.